Changed project structure to share code across projects
The JavaScript code lives now in hector-js which references hector-base.
hector-web depends on hector-js and hector-base.

hector-base contains general utility code like TextOutput.

This new structure allows us to introduce hector-macro in the
next step which can be used by hector-web and depend on hector-js.
This will lead to a js-macro that transforms a Scala AST into a
JavaScript AST.